requirements.txt
flask
firebase_admin
numpy
flask-socketio
gevent
eventlet 